garbage_classification/python/requirements.txt
tensorflow==2.10.0
numpy<2.0.0
Pillow
opencv-python
matplotlib>=3.7.0
scikit-learn>=1.3.0
tf-keras
tensorflow-model-optimization
